backend/requirements.txt
Flask==3.1.0
snowflake-connector-python==3.12.4
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.1.0